[broken] Non-test mode is compiling!
