+++ Cargo.toml
@@ -12,7 +12,8 @@ categories = ["development-tools"]
 resolver = "2"
 
 [workspace]
-members = [".", "fsm", "test_utils", "sdk-core-protos", "temporal-sdk"]
+members = [".", "temporal-client", "core-api", "fsm", "test_utils",
+           "sdk-core-protos", "temporal-sdk"]
 
 [lib]
 
@@ -20,7 +21,6 @@ members = [".", "fsm", "test_utils", "sdk-core-protos", "temporal-sdk"]
 anyhow = "1.0"
 arc-swap = "1.3"
 async-trait = "0.1"
-backoff = "0.3.0"
 base64 = "0.13"
 crossbeam = "0.8"
 dashmap = "4.0"
@@ -28,7 +28,6 @@ derive_builder = "0.10"
 derive_more = "0.99"
 enum_dispatch = "0.3"
 futures = "0.3"
-futures-retry = "0.6.0"
 http = "0.2"
 hyper = "0.14"
 itertools = "0.10"
@@ -61,8 +60,16 @@ url = "2.2"
 uuid = { version = "0.8.2", features = ["v4"] }
 
 # 1st party local deps
+[dependencies.temporal-sdk-core-api]
+path = "core-api"
+version = "0.1"
+
 [dependencies.temporal-sdk-core-protos]
 path = "sdk-core-protos"
+version = "0.1"
+
+[dependencies.temporal-client]
+path = "temporal-client"
 version = "0.1"
 
 [dependencies.rustfsm]
