+++ core-api/Cargo.toml
@@ -1,23 +1,23 @@
 [package]
-name = "test_utils"
+name = "temporal-sdk-core-api"
 version = "0.1.0"
-authors = ["Spencer Judge <spencer@temporal.io>"]
 edition = "2021"
-resolver = "2"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0"
 async-trait = "0.1"
-base64 = "0.13"
-futures = "0.3"
+derive_builder = "0.10"
 log = "0.4"
-rand = "0.8"
-temporal-sdk-core = { path = ".." }
-temporal-sdk-core-api = { path = "../core-api" }
-temporal-sdk = { path = "../temporal-sdk" }
-url = "2.2"
+prost-types = "0.9"
+thiserror = "1.0"
+tonic = "0.6"
 
 [dependencies.temporal-sdk-core-protos]
 path = "../sdk-core-protos"
+version = "0.1"
+
+[dependencies.temporal-client]
+path = "../temporal-client"
 version = "0.1"
