+++ sdk-core-protos/src/lib.rs
@@ -4,6 +4,10 @@
 
 pub mod utilities;
 
+mod task_token;
+
+pub use task_token::TaskToken;
+
 #[allow(clippy::large_enum_variant)]
 // I'd prefer not to do this, but there are some generated things that just don't need it.
 #[allow(missing_docs)]
