+++ src/lib.rs
@@ -10,14 +10,11 @@ pub extern crate assert_matches;
 #[macro_use]
 extern crate tracing;
 
-pub mod errors;
-
 mod log_export;
 mod pending_activations;
 mod pollers;
 mod protosext;
 pub(crate) mod retry_logic;
-pub(crate) mod task_token;
 pub(crate) mod telemetry;
 mod worker;
 mod workflow;
@@ -28,21 +25,18 @@ mod core_tests;
 #[macro_use]
 mod test_help;
 
-pub use log_export::CoreLog;
+pub(crate) use temporal_sdk_core_api::errors;
+
 pub use pollers::{
     ClientTlsConfig, RetryConfig, RetryGateway, ServerGateway, ServerGatewayApis,
     ServerGatewayOptions, ServerGatewayOptionsBuilder, TlsConfig,
 };
-pub use task_token::TaskToken;
 pub use telemetry::{TelemetryOptions, TelemetryOptionsBuilder};
+pub use temporal_sdk_core_protos::TaskToken;
 pub use url::Url;
 pub use worker::{WorkerConfig, WorkerConfigBuilder};
 
 use crate::{
-    errors::{
-        ActivityHeartbeatError, CompleteActivityError, CompleteWfError, CoreInitError,
-        PollActivityError, PollWfError, WorkerRegistrationError,
-    },
     pollers::GatewayRef,
     telemetry::{fetch_global_buffered_logs, telemetry_init},
     worker::{Worker, WorkerDispatcher},
@@ -53,6 +47,13 @@ use std::{
         atomic::{AtomicBool, Ordering},
         Arc,
     },
+};
+use temporal_sdk_core_api::{
+    errors::{
+        CompleteActivityError, CompleteWfError, CoreInitError, PollActivityError, PollWfError,
+        WorkerRegistrationError,
+    },
+    Core, CoreLog,
 };
 use temporal_sdk_core_protos::coresdk::{
     activity_task::ActivityTask, workflow_activation::WfActivation,
@@ -68,123 +69,6 @@ lazy_static::lazy_static! {
     static ref PROCCESS_UNIQ_ID: String = {
         uuid::Uuid::new_v4().to_simple().to_string()
     };
-}
-
-/// This trait is the primary way by which language specific SDKs interact with the core SDK. It is
-/// expected that only one instance of an implementation will exist for the lifetime of the
-/// worker(s) using it.
-#[async_trait::async_trait]
-pub trait Core: Send + Sync {
-    /// Register a worker with core. Workers poll on a specific task queue, and when calling core's
-    /// poll functions, you must provide a task queue name. If there was already a worker registered
-    /// with the same task queue name, it will be shut down and a new one will be created.
-    async fn register_worker(&self, config: WorkerConfig) -> Result<(), WorkerRegistrationError>;
-
-    /// Ask the core for some work, returning a [WfActivation]. It is then the language SDK's
-    /// responsibility to call the appropriate workflow code with the provided inputs. Blocks
-    /// indefinitely until such work is available or [Core::shutdown] is called.
-    ///
-    /// The returned activation is guaranteed to be for the same task queue / worker which was
-    /// provided as the `task_queue` argument.
-    ///
-    /// It is important to understand that all activations must be responded to. There can only
-    /// be one outstanding activation for a particular run of a workflow at any time. If an
-    /// activation is not responded to, it will cause that workflow to become stuck forever.
-    ///
-    /// Activations that contain only a `remove_from_cache` job should not cause the workflow code
-    /// to be invoked and may be responded to with an empty command list. Eviction jobs may also
-    /// appear with other jobs, but will always appear last in the job list. In this case it is
-    /// expected that the workflow code will be invoked, and the response produced as normal, but
-    /// the caller should evict the run after doing so.
-    ///
-    /// It is rarely a good idea to call poll concurrently. It handles polling the server
-    /// concurrently internally.
-    ///
-    /// TODO: Examples
-    async fn poll_workflow_activation(&self, task_queue: &str)
-        -> Result<WfActivation, PollWfError>;
-
-    /// Ask the core for some work, returning an [ActivityTask]. It is then the language SDK's
-    /// responsibility to call the appropriate activity code with the provided inputs. Blocks
-    /// indefinitely until such work is available or [Core::shutdown] is called.
-    ///
-    /// The returned activation is guaranteed to be for the same task queue / worker which was
-    /// provided as the `task_queue` argument.
-    ///
-    /// It is rarely a good idea to call poll concurrently. It handles polling the server
-    /// concurrently internally.
-    ///
-    /// TODO: Examples
-    async fn poll_activity_task(&self, task_queue: &str)
-        -> Result<ActivityTask, PollActivityError>;
-
-    /// Tell the core that a workflow activation has completed. May be freely called concurrently.
-    async fn complete_workflow_activation(
-        &self,
-        completion: WfActivationCompletion,
-    ) -> Result<(), CompleteWfError>;
-
-    /// Tell the core that an activity has finished executing. May be freely called concurrently.
-    async fn complete_activity_task(
-        &self,
-        completion: ActivityTaskCompletion,
-    ) -> Result<(), CompleteActivityError>;
-
-    /// Notify workflow that an activity is still alive. Long running activities that take longer
-    /// than `activity_heartbeat_timeout` to finish must call this function in order to report
-    /// progress, otherwise the activity will timeout and a new attempt will be scheduled.
-    ///
-    /// The first heartbeat request will be sent immediately, subsequent rapid calls to this
-    /// function will result in heartbeat requests being aggregated and the last one received during
-    /// the aggregation period will be sent to the server, where that period is defined as half the
-    /// heartbeat timeout.
-    ///
-    /// Unlike java/go SDKs we do not return cancellation status as part of heartbeat response and
-    /// instead send it as a separate activity task to the lang, decoupling heartbeat and
-    /// cancellation processing.
-    ///
-    /// For now activity still need to send heartbeats if they want to receive cancellation
-    /// requests. In the future we will change this and will dispatch cancellations more
-    /// proactively. Note that this function does not block on the server call and returns
-    /// immediately. Underlying validation errors are swallowed and logged, this has been agreed to
-    /// be optimal behavior for the user as we don't want to break activity execution due to badly
-    /// configured heartbeat options.
-    fn record_activity_heartbeat(&self, details: ActivityHeartbeat);
-
-    /// Request that a workflow be evicted by its run id. This will generate a workflow activation
-    /// with the eviction job inside it to be eventually returned by
-    /// [Core::poll_workflow_activation]. If the workflow had any existing outstanding activations,
-    /// such activations are invalidated and subsequent completions of them will do nothing and log
-    /// a warning.
-    fn request_workflow_eviction(&self, task_queue: &str, run_id: &str);
-
-    /// Returns core's instance of the [ServerGatewayApis] implementor it is using.
-    fn server_gateway(&self) -> Arc<dyn ServerGatewayApis + Send + Sync>;
-
-    /// Initiates async shutdown procedure, eventually ceases all polling of the server and shuts
-    /// down all registered workers. [Core::poll_workflow_activation] should be called until it
-    /// returns [PollWfError::ShutDown] to ensure that any workflows which are still undergoing
-    /// replay have an opportunity to finish. This means that the lang sdk will need to call
-    /// [Core::complete_workflow_activation] for those workflows until they are done. At that point,
-    /// the lang SDK can end the process, or drop the [Core] instance, which will close the
-    /// connection.
-    async fn shutdown(&self);
-
-    /// Shut down a specific worker. Will cease all polling on the task queue and future attempts
-    /// to poll that queue will return [PollWfError::NoWorkerForQueue].
-    async fn shutdown_worker(&self, task_queue: &str);
-
-    /// Retrieve options that were passed in when initializing core
-    fn get_init_options(&self) -> &CoreInitOptions;
-
-    /// Core buffers logs that should be shuttled over to lang so that they may be rendered with
-    /// the user's desired logging library. Use this function to grab the most recent buffered logs
-    /// since the last time it was called. A fixed number of such logs are retained at maximum, with
-    /// the oldest being dropped when full.
-    ///
-    /// Returns the list of logs from oldest to newest. Returns an empty vec if the feature is not
-    /// configured.
-    fn fetch_buffered_logs(&self) -> Vec<CoreLog>;
 }
 
 /// Holds various configuration information required to call [init]
@@ -328,10 +212,6 @@ impl Core for CoreSDK {
         self.workers.shutdown_one(task_queue).await;
     }
 
-    fn get_init_options(&self) -> &CoreInitOptions {
-        &self.init_options
-    }
-
     fn fetch_buffered_logs(&self) -> Vec<CoreLog> {
         fetch_global_buffered_logs()
     }
@@ -402,4 +282,42 @@ impl CoreSDK {
 enum WorkerLookupErr {
     Shutdown(String),
     NoWorker(String),
+}
+
+impl From<WorkerLookupErr> for PollWfError {
+    fn from(e: WorkerLookupErr) -> Self {
+        match e {
+            WorkerLookupErr::Shutdown(_) => Self::ShutDown,
+            WorkerLookupErr::NoWorker(s) => Self::NoWorkerForQueue(s),
+        }
+    }
+}
+
+impl From<WorkerLookupErr> for PollActivityError {
+    fn from(e: WorkerLookupErr) -> Self {
+        match e {
+            WorkerLookupErr::Shutdown(_) => Self::ShutDown,
+            WorkerLookupErr::NoWorker(s) => Self::NoWorkerForQueue(s),
+        }
+    }
+}
+
+impl From<WorkerLookupErr> for CompleteWfError {
+    fn from(e: WorkerLookupErr) -> Self {
+        match e {
+            WorkerLookupErr::Shutdown(s) | WorkerLookupErr::NoWorker(s) => {
+                Self::NoWorkerForQueue(s)
+            }
+        }
+    }
+}
+
+impl From<WorkerLookupErr> for CompleteActivityError {
+    fn from(e: WorkerLookupErr) -> Self {
+        match e {
+            WorkerLookupErr::Shutdown(s) | WorkerLookupErr::NoWorker(s) => {
+                Self::NoWorkerForQueue(s)
+            }
+        }
+    }
 }
