+++ src/pollers/mod.rs
@@ -1,19 +1,17 @@
-mod gateway;
 mod poll_buffer;
-mod retry;
 
-#[cfg(test)]
-pub use gateway::{MockManualGateway, MockServerGatewayApis};
+// #[cfg(test)]
+// pub use gateway::{MockManualGateway, MockServerGatewayApis};
 
-pub use gateway::{
-    ClientTlsConfig, GatewayRef, RetryConfig, ServerGateway, ServerGatewayApis,
-    ServerGatewayOptions, ServerGatewayOptionsBuilder, TlsConfig,
-};
 pub use poll_buffer::{
     new_activity_task_buffer, new_workflow_task_buffer, PollActivityTaskBuffer,
     PollWorkflowTaskBuffer, WorkflowTaskPoller,
 };
-pub use retry::RetryGateway;
+use std::{ops::Deref, sync::Arc};
+pub use temporal_client::{
+    ClientTlsConfig, RetryConfig, RetryGateway, ServerGateway, ServerGatewayApis,
+    ServerGatewayOptions, ServerGatewayOptionsBuilder, TlsConfig,
+};
 
 use temporal_sdk_core_protos::temporal::api::workflowservice::v1::{
     PollActivityTaskQueueResponse, PollWorkflowTaskQueueResponse,
@@ -21,20 +19,30 @@ use temporal_sdk_core_protos::temporal::api::workflowservice::v1::{
 
 #[cfg(test)]
 use futures::Future;
-use tonic::Code;
 
 pub type Result<T, E = tonic::Status> = std::result::Result<T, E>;
 
-/// List of gRPC error codes that client will retry.
-pub const RETRYABLE_ERROR_CODES: [Code; 7] = [
-    Code::DataLoss,
-    Code::Internal,
-    Code::Unknown,
-    Code::ResourceExhausted,
-    Code::Aborted,
-    Code::OutOfRange,
-    Code::Unavailable,
-];
+pub struct GatewayRef {
+    pub gw: Arc<dyn ServerGatewayApis + Send + Sync>,
+    pub options: ServerGatewayOptions,
+}
+
+impl GatewayRef {
+    pub fn new<SG: ServerGatewayApis + Send + Sync + 'static>(
+        gw: Arc<SG>,
+        options: ServerGatewayOptions,
+    ) -> Self {
+        Self { gw, options }
+    }
+}
+
+impl Deref for GatewayRef {
+    type Target = dyn ServerGatewayApis + Send + Sync;
+
+    fn deref(&self) -> &Self::Target {
+        self.gw.as_ref()
+    }
+}
 
 /// A trait for things that poll the server. Hides complexity of concurrent polling or polling
 /// on sticky/nonsticky queues simultaneously.
