+++ src/worker/activities/local_activities.rs
@@ -1,4 +1,4 @@
-use crate::{protosext::ValidScheduleLA, retry_logic::RetryPolicyExt, task_token::TaskToken};
+use crate::{protosext::ValidScheduleLA, retry_logic::RetryPolicyExt, TaskToken};
 use parking_lot::Mutex;
 use std::{
     collections::HashMap,
