+++ temporal-sdk/Cargo.toml
@@ -26,3 +26,7 @@ version = "0.1"
 [dependencies.temporal-sdk-core-protos]
 path = "../sdk-core-protos"
 version = "0.1"
+
+[dependencies.temporal-sdk-core-api]
+path = "../core-api"
+version = "0.1"
