+++ test_utils/src/lib.rs
@@ -4,10 +4,11 @@ use rand::{distributions::Standard, Rng};
 use std::{convert::TryFrom, env, future::Future, net::SocketAddr, sync::Arc, time::Duration};
 use temporal_sdk::TestRustWorker;
 use temporal_sdk_core::{
-    Core, CoreInitOptions, CoreInitOptionsBuilder, ServerGatewayApis, ServerGatewayOptions,
+    CoreInitOptions, CoreInitOptionsBuilder, ServerGatewayApis, ServerGatewayOptions,
     ServerGatewayOptionsBuilder, TelemetryOptions, TelemetryOptionsBuilder, WorkerConfig,
     WorkerConfigBuilder,
 };
+use temporal_sdk_core_api::Core;
 use temporal_sdk_core_protos::coresdk::{
     workflow_commands::{
         workflow_command, ActivityCancellationType, CompleteWorkflowExecution, ScheduleActivity,
